% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.extract.time.series}
\alias{dp.extract.time.series}
\title{Helper function for extracting a single row of data with one value per year}
\usage{
dp.extract.time.series(
  dp.raw,
  direction = "wide",
  first.year,
  final.year,
  tag,
  offset
)
}
\description{
Helper function for extracting a single row of data with one value per year
}
