% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.art.adjustment.flag}
\alias{dp.inputs.adult.art.adjustment.flag}
\title{Check whether adult ART inputs were adjusted}
\usage{
dp.inputs.adult.art.adjustment.flag(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
TRUE if adult ART inputs were adjusted, FALSE otherwise
}
\description{
Check whether adult ART inputs were adjusted
}
\section{Details}{


  The magnitude of adjustments made to ART inputs can be accessed using
  \code{dp.inputs.adult.art.adjustment.value}.
}

