% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.art.allocation}
\alias{dp.inputs.adult.art.allocation}
\title{Check parameters used to determine ART initiation timing}
\usage{
dp.inputs.adult.art.allocation(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
A list with named elements 'Method' and 'Weight'
}
\description{
Check parameters used to determine ART initiation timing
}
\section{Details}{


  Spectrum uses one of four methods to determine how new ART patients are
  drawn from the ART-eligible population by CD4 category:
  \enumerate{
  \item{Method=1: ART initiations are proportional to the expected number of HIV-related deaths in each CD4 category.}
  \item{Method=2: ART initiations are proportional to the number of people in each CD4 category.}
  \item{Method=3: ART initiations are a weighted combination of methods 1 and 2.}
  \item{Method=4: ART initiations are drawn from the lowest CD4 categories first.}
  }
  The return value element "Weight"=(w1, w2) is a two-element vector that
  describes the weight placed on method 1 (w1) and method 2 (w2) when using
  methods 1, 2, or 3. Spectrum ignores these weights when method 4 is used.
}

