% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.art.monthly}
\alias{dp.inputs.adult.art.monthly}
\title{Get numbers of adults on ART by month during 2020-2021}
\usage{
dp.inputs.adult.art.monthly(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A data frame.
}
\description{
Get the number of adults on ART during 2020-2021, stratified by sex.
}
