% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.art}
\alias{dp.inputs.adult.art}
\title{Get Spectrum adult ART inputs}
\usage{
dp.inputs.adult.art(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get input numbers or percentages of adults on ART entered into Spectrum by
sex and year in long or wide format
}
\section{Details}{


  Adult treatment data can be entered into Spectrum as numbers or
  percentages. These units can vary from year to year. When
  \code{direction="wide"}, the return value will include rows for numbers and
  for percentages, with percentages missing in years where numbers were
  entered or vice-versa. When \code{direction="long"}, the data frame will only
  include rows for whichever unit was entered into Spectrum in any year.
}

