% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.cd4.progression}
\alias{dp.inputs.adult.cd4.progression}
\title{Get the adult HIV disease progression rates}
\usage{
dp.inputs.adult.cd4.progression(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
adult HIV disease progression rates
}
\description{
Get the adult HIV disease progression rates
}
