% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.hiv.mortality.art}
\alias{dp.inputs.adult.hiv.mortality.art}
\alias{dp.inputs.adult.hiv.mortality.art.trend}
\alias{dp.inputs.adult.hiv.mortality.art.scale}
\title{HIV-related mortality rates among adults on ART}
\usage{
dp.inputs.adult.hiv.mortality.art(dp.raw, direction = "wide")

dp.inputs.adult.hiv.mortality.art.trend(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.inputs.adult.hiv.mortality.art.scale(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
\code{dp.inputs.adult.hiv.mortality.art} returns a data frame.

\code{dp.inputs.adult.hiv.mortality.art.trend} returns a data frame.

\code{dp.inputs.adult.hiv.mortality.art.scale} returns scale factor.
}
\description{
HIV-related mortality rate inputs to Spectrum are stratified by age, sex, CD4
category, time on ART, and calendar year. Spectrum users may adjust a
separate on-ART mortality scale factor to calibrate deaths on ART to vital
registration data. Baseline rates by age, sex, CD4 and ART duration are
accessed separately from mortality rate ratios over time and scale factors.
Functions to access each of these are described below.
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.adult.hiv.mortality.art()}: Baseline HIV-related mortality
rates on ART by sex, age, CD4 category, and time on ART.

\item \code{dp.inputs.adult.hiv.mortality.art.trend()}: On-ART mortality rate ratios by year and ART duration.

\item \code{dp.inputs.adult.hiv.mortality.art.scale()}: Scale factor applied to on-ART mortality rates.

}}
