% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.hiv.mortality.art.scale}
\alias{dp.inputs.adult.hiv.mortality.art.scale}
\title{Get the HIV-related mortality rate ratio among adults on ART}
\usage{
dp.inputs.adult.hiv.mortality.art.scale(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
the mortality multiplier
}
\description{
Get the HIV-related mortality rate ratio among adults on ART. This is a
scalar multiplier applied on top of mortality rate ratio trends by year and
time on ART.
}
