% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.hiv.mortality.custom}
\alias{dp.inputs.adult.hiv.mortality.custom}
\alias{dp.inputs.adult.hiv.mortality.art.custom}
\title{AIM advanced option configuration}
\usage{
dp.inputs.adult.hiv.mortality.custom(dp.raw, direction = "wide")

dp.inputs.adult.hiv.mortality.art.custom(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
TRUE if the parameter is unlocked for configuration, FALSE otherwise.
}
\description{
Check if advanced options in AIM are unlocked for customization.
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.adult.hiv.mortality.custom()}: Mortality rates among HIV+ adults not on ART

\item \code{dp.inputs.adult.hiv.mortality.art.custom()}: Mortality rates among HIV+ adults on ART

}}
