% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.hiv.mortality.nonaids}
\alias{dp.inputs.adult.hiv.mortality.nonaids}
\title{Get rates of non-AIDS excess mortality among adults with HIV}
\usage{
dp.inputs.adult.hiv.mortality.nonaids(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A data frame.
}
\description{
Get rates of non-AIDS excess mortality among adults with HIV
}
