% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.hiv.mortality.off.art}
\alias{dp.inputs.adult.hiv.mortality.off.art}
\title{Get HIV-related mortality rates among adults not on ART}
\usage{
dp.inputs.adult.hiv.mortality.off.art(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
adult HIV-related mortality rates off ART
}
\description{
Get HIV-related mortality rates among adults not on ART
}
