% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.hiv.mortality.region}
\alias{dp.inputs.adult.hiv.mortality.region}
\alias{dp.inputs.adult.hiv.mortality.art.region}
\title{AIM advanced options regional configuration}
\usage{
dp.inputs.adult.hiv.mortality.region(dp.raw, direction = "wide")

dp.inputs.adult.hiv.mortality.art.region(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
The region name as a factor (see "Details" below for factor levels)
}
\description{
Get the region selected for advanced options parameter values in Spectrum
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.adult.hiv.mortality.region()}: Mortality rates among HIV+ adults not on ART

\item \code{dp.inputs.adult.hiv.mortality.art.region()}: Mortality rates among HIV+ adults on ART

}}
\section{Details}{


Spectrum options are specified for one of several regions:
\enumerate{
\item{Asia}
\item{Central Africa}
\item{Developed Countries}
\item{East Africa}
\item{Eastern Europe}
\item{Latin America and Caribbean}
\item{North Africa Middle East}
\item{Southern Africa}
\item{West Africa}
}
}

