% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.initial.cd4}
\alias{dp.inputs.adult.initial.cd4}
\title{Get the initial distribution of newly-infected adults by CD4 cell count category}
\usage{
dp.inputs.adult.initial.cd4(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
the initial CD4 category distribution for newly-infected adults
}
\description{
Get the initial distribution of newly-infected adults by CD4 cell count category
}
