% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.adult.ltfu.monthly}
\alias{dp.inputs.adult.ltfu.monthly}
\title{Get adult ART loss to follow-up inputs by month during 2020-2021}
\usage{
dp.inputs.adult.ltfu.monthly(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A data frame.
}
\description{
Get the input monthly percentages of adults who were lost to ART follow-up during 2020-2021.
}
