% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.art.by.age}
\alias{dp.inputs.art.by.age}
\title{Get Spectrum ART by age inputs}
\usage{
dp.inputs.art.by.age(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get input numbers of people on ART entered into Spectrum by sex, age and year
in long or wide format. These data may be entered by five-year age group, or
in age groups specified for UNAIDS Global AIDS Monitoring. The latter
stratifies ages 25+ into two age groups, 25-49 and 50+.
}
