% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.art.transmission.reduction}
\alias{dp.inputs.art.transmission.reduction}
\title{Get the effect of ART on HIV transmission}
\usage{
dp.inputs.art.transmission.reduction(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
a number
}
\description{
Return the reduction in HIV transmission on ART. Spectrum does not use this
in its calculations, but passes it to EPP for use in incidence estimation.
}
