% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.child.art}
\alias{dp.inputs.child.art}
\title{Get Spectrum child HIV treatment inputs}
\usage{
dp.inputs.child.art(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get input numbers or percentages of children on antiretroviral therapy (ART)
or cotrimoxazole (CTX) prophylaxis entered into Spectrum by year and age in
long or wide format
}
\section{Details}{


  Child treatment data can be entered into Spectrum as numbers or percentages
  for ages 0-14. Further, ART numbers can be entered by five-year age group
  (0-4, 5-9, or 10-14) instead. These units and age resolutions can vary from
  year to year. When \code{direction="wide"}, the return value will include
  rows for numbers and for percentages, with percentages missing in years
  where numbers were entered or vice-versa. When \code{direction="long"},
  unused combinations of age group, treatment, and unit are removed before
  returning.
}

