% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.covid19.enabled}
\alias{dp.inputs.covid19.enabled}
\title{Check if COVID-19 inputs are enabled}
\usage{
dp.inputs.covid19.enabled(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
TRUE if the toggle is enabled, FALSE otherwise.
}
\description{
Countries can toggle entry of COVID-19 deaths as a configuration option. This
function checks whether that toggle is turned on.
}
