% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.csavr.data.options}
\alias{dp.inputs.csavr.data.options}
\title{Check which data were selected for CSAVR incidence estimation}
\usage{
dp.inputs.csavr.data.options(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
The incidence model name as a factor (see "Details" below for factor
  levels)
}
\description{
Check which data were selected for CSAVR incidence estimation
}
\section{Details}{


  The CSAVR fitting tool can estimate incidence using three different kinds
  of data
  \enumerate{
  \item{New HIV diagnoses}
  \item{HIV-related deaths}
  \item{CD4 cell counts at diagnosis (as of 2022)}
  }

  Use \code{dp.inputs.incidence.model()} to check if CSAVR was used to
  estimate incidence; use \code{dp.inputs.incidence()} to get the incidence
  estimate itself.

  Since CD4 counts at diagnoses were reintroduced in Spectrum for the 2022
  round of UNAIDS HIV estimates, this function may not work correctly with
  files produced with older versions of Spectrum.
}

