% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.csavr.deaths.source}
\alias{dp.inputs.csavr.deaths.source}
\alias{dp.inputs.csavr.deaths.source.names}
\alias{dp.inputs.csavr.deaths}
\alias{dp.inputs.csavr.deaths.sex}
\alias{dp.inputs.csavr.deaths.sex.age}
\title{Get input numbers of AIDS deaths used by CSAVR to estimate HIV incidence}
\usage{
dp.inputs.csavr.deaths.source(dp.raw, direction = "wide")

dp.inputs.csavr.deaths.source.names(dp.raw, direction = "wide")

dp.inputs.csavr.deaths(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.inputs.csavr.deaths.sex(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.inputs.csavr.deaths.sex.age(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
a data frame
}
\description{
Spectrum's Case Surveillance and Vital Registration (CSAVR) tool uses case
surveillance data on new HIV diagnoses and vital registration data on AIDS
deaths in adults to estimate HIV incidence. AIDS deaths can be entered
overall, by sex, or by sex and age. Countries may use AIDS deaths data from
their vital registration system or AIDS deaths estimates produced by the
Institute of Health Metrics and Evaluation (IHME) in CSAVR. Current Spectrum
versions allow users to enter data from up to three different sources, then
select which source would be used for fitting. Older versions of Spectrum
only allowed users to enter one source of data.
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.csavr.deaths.source()}: The data source selected for use in
CSAVR.

\item \code{dp.inputs.csavr.deaths.source.names()}: User-specified name for each source of AIDS deaths data entered into CSAVR.

\item \code{dp.inputs.csavr.deaths()}: Adult AIDS deaths by year and data source.

\item \code{dp.inputs.csavr.deaths.sex()}: Adult AIDS deaths by year, sex, and data source.

\item \code{dp.inputs.csavr.deaths.sex.age()}: Adult AIDS deaths by year, sex, age, and data source.

}}
