% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.csavr.irr.options}
\alias{dp.inputs.csavr.irr.options}
\title{Check if incidence rate ratios (IRRs) by sex or age were estimated while fitting CSAVR}
\usage{
dp.inputs.csavr.irr.options(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
A data frame of TRUE/FALSE variables indicating whether IRR fitting by sex or
age was enabled. These options are selected separately for each CSAVR incidence model.
}
\description{
Check if incidence rate ratios (IRRs) by sex or age were estimated while fitting CSAVR
}
\section{Limitations}{


  Users may change IRR settings in Spectrum after they have fitted an incidence model. As
  a result, the IRR option indicators may not always correctly indicate whether IRRs were
  estimated during fitting.
}

