% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.csavr.model}
\alias{dp.inputs.csavr.model}
\title{Get the model used to estimate incidence in CSAVR}
\usage{
dp.inputs.csavr.model(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
The incidence model name as a factor (see "Details" below for factor
  levels)
}
\description{
Get the model used to estimate incidence in CSAVR
}
\section{Details}{


  CSAVR users select one of four curve types to model incidence:
  \enumerate{
  \item{None - no curve selected}
  \item{Single logistic}
  \item{Double logistic}
  \item{Splines}
  \item{rLogistic}
  }

  Use \code{dp.inputs.incidence.model()} to check if CSAVR was used to
  estimate incidence; use \code{dp.inputs.incidence()} to get the incidence
  estimate itself.
}

