% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.e0.default}
\alias{dp.inputs.e0.default}
\title{Get the default input life expectancy at birth}
\usage{
dp.inputs.e0.default(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get the default input life expectancy at birth (e0) by year and sex. These
life expectancy trends typically come from the latest UN Population Division
World Population Prospects revision. These may differ from the user-editable
life expectancy used in Spectrum calculation, which can be accessed using
\code{dp.inputs.e0}.
}
