% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.epp.adjustment.cap}
\alias{dp.inputs.epp.adjustment.cap}
\title{Get the maximum amount that AIM can adjust incidence estimates from EPP}
\usage{
dp.inputs.epp.adjustment.cap(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
the maximum allowed adjustment.
}
\description{
EPP hands AIM estimated HIV incidence and prevalence trends. EPP and AIM
have different structures, so EPP's incidence may produce a somewhat different prevalence trend
than EPP provides. EPP estimates its trends using HIV prevalence data, so
AIM may adjust the incidence it gets from EPP to better match EPP's prevalence
estimate, subject to a user-specified maximum allowed adjustment. This
function returns that maximum. The minimum adjustment is equal to the reciprocal
of the maximum.
}
