% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.epp.adjustment.enabled}
\alias{dp.inputs.epp.adjustment.enabled}
\title{Check if AIM is allowed to adjust HIV incidence from EPP}
\usage{
dp.inputs.epp.adjustment.enabled(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
TRUE if adjustments is allowed, FALSE otherwise
}
\description{
Check if AIM is allowed to adjust HIV incidence from EPP
}
\section{Details}{


  The cap on adjustments allowed can be accessed using
  \code{dp.inputs.epp.adjustment.cap}. See the documentation of that function
  for more details on the rationale for adjustment.
}

