% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.epp.epidemic.first.year}
\alias{dp.inputs.epp.epidemic.first.year}
\title{Get the first year of HIV incidence estimated by EPP}
\usage{
dp.inputs.epp.epidemic.first.year(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
an integer-valued year
}
\description{
Get the first year of HIV incidence estimated by EPP
}
