% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.external.pop}
\alias{dp.inputs.external.pop}
\title{Get the external population inputs used to calculate population adjustments}
\usage{
dp.inputs.external.pop(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Spectrum users may enter externally-produced population targets by sex, age,
and year that can be used to align Spectrum population outputs with a
reference population projection. \code{dp.inputs.external.pop} returns those
population targets in long or wide format.
}
\section{Details}{


  Use of an external population file is optional and rare in Spectrum. Use
  \code{dp.inputs.use.external.pop} to check whether to check whether an
  external population was specified. Note that \code{dp.inputs.external.pop}
  will return population sizes of 0 for all years, sexes and ages if no
  external population was specified.
}

