% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.final.year}
\alias{dp.inputs.final.year}
\title{Get the final year of a Spectrum projection}
\usage{
dp.inputs.final.year(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
the final year of the projection.
}
\description{
Get the final year of a Spectrum projection
}
