% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.first.year}
\alias{dp.inputs.first.year}
\alias{dp.inputs.final.year}
\title{Spectrum projection time span}
\usage{
dp.inputs.first.year(dp.raw, direction = "wide")

dp.inputs.final.year(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
the requested year
}
\description{
Spectrum projection time span
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.first.year()}: First year of the projection

\item \code{dp.inputs.final.year()}: Final year of the projection

}}
