% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.hiv.breastfeeding}
\alias{dp.inputs.hiv.breastfeeding}
\title{Get breastfeeding patterns in mothers with HIV}
\usage{
dp.inputs.hiv.breastfeeding(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Breastfeeding practices in mothers with HIV are specified in Spectrum as the
percentage of mothers who are not breastfeeding by year, child's age, and
whether the mother is receiving ARVs to prevent vertical HIV transmission. Child's age
is specified by two month intervals up to 36 months.
}
