% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.hiv.frr.cd4}
\alias{dp.inputs.hiv.frr.cd4}
\title{Get HIV-related fertility adjustments by CD4 cell count}
\usage{
dp.inputs.hiv.frr.cd4(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{Ignored; included for compatibility with similar functions.}

\item{final.year}{Ignored; included for compatibility with similar functions.}
}
\value{
A data frame.
}
\description{
Get HIV-related fertility adjustments by CD4 cell count category for women
with untreated HIV
}
