% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.hiv.frr.location}
\alias{dp.inputs.hiv.frr.location}
\title{Get the HIV-related fertility local adjustment factor}
\usage{
dp.inputs.hiv.frr.location(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}

\item{first.year}{Ignored; included for compatibility with similar functions.}

\item{final.year}{Ignored; included for compatibility with similar functions.}
}
\value{
the HIV-related fertility local adjustment factor
}
\description{
Get the HIV-related fertility local adjustment factor
}
