% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.incidence.model}
\alias{dp.inputs.incidence.model}
\title{Get the model used to estimate incidence in a Spectrum projection}
\usage{
dp.inputs.incidence.model(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
The incidence model name as a factor (see "Details" below for factor levels)
}
\description{
Get the model used to estimate incidence in a Spectrum projection
}
\section{Details}{


  Spectrum can take HIV incidence from several different models as inputs:
  \enumerate{
  \item{Direct - manually entered incidence values by year}
  \item{EPP - the Estimation and Projection Package}
  \item{AEM - the AIDS Epidemic Model}
  \item{CSAVR - the Case Surveillance and Vital Registration tool}
  \item{Mortality - incidence calibrated to mortality data}
  \item{ECDC - The European Centre for Disease Prevention and Control's HIV Modelling Tool}
  }

  \code{dp.inputs.incidence.model()} returns the name of the model used as a factor. Use
  \code{dp.inputs.incidence()} to get the incidence trend itself.
}

