% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.irr.custom}
\alias{dp.inputs.irr.custom}
\title{Check if Spectrum used custom incidence rate ratios by age or sex}
\usage{
dp.inputs.irr.custom(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
TRUE if custom incidence rate ratios were used, FALSE otherwise
}
\description{
Check if Spectrum used custom incidence rate ratios by age or sex
}
