% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.irr.sex.from.epp}
\alias{dp.inputs.irr.sex.from.epp}
\title{Check if Spectrum used incidence rate ratios by sex estimated by EPP}
\usage{
dp.inputs.irr.sex.from.epp(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
TRUE if sex ratios from EPP were selected, FALSE otherwise
}
\description{
Check if Spectrum used incidence rate ratios by sex estimated by EPP
}
