% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.kos.source}
\alias{dp.inputs.kos.source}
\title{Get the source indicated for number who know their HIV+ status}
\usage{
dp.inputs.kos.source(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
The knowledge of status (KoS) source as a factor (see "Details" below
  for factor levels)
}
\description{
Get the source indicated for number who know their HIV+ status
}
\section{Details}{


  Spectrum users can choose one of five sources for numbers of people living
  with HIV who know their HIV status:
  \enumerate{
  \item{Case reports}
  \item{Shiny90}
  \item{CSAVR - the Case Surveillance and Vital Registration tool}
  \item{ECDC - The European Centre for Disease Prevention and Control's HIV Modelling Tool}
  \item{Direct - manually entered values by year}
  }

Note that users may change this selection after entering knowledge of status data into Spectrum.
\code{dp.inputs.kos.data()} returns the input numbers of people who know their status.
}

