% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.life.table}
\alias{dp.inputs.life.table}
\title{Get the input life table selection}
\usage{
dp.inputs.life.table(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
The life table name as a factor
}
\description{
Spectrum users may select a life table from a list of model life tables, or use a
country-specific life table based on the latest World Population Prospects
revision. This function indicates which selection the user made.
}
\section{Details}{


  \code{dp.inputs.life.table} just indicates which life table was selected.
  Life table indicators can be accessed separately: \code{dp.inputs.surv}
  returns survival ratios, and \code{dp.inputs.e0} returns life expectancy at
  birth. To keep projection file sizes small, most other common life table
  indicators are not saved.

  Supported life tables include:
  \enumerate{
  \item{Coale-Demeny West}
  \item{Coale-Demeny North}
  \item{Coale-Demeny East}
  \item{Coale-Demeny South}
  \item{UN General}
  \item{UN Chile}
  \item{UN South Asia}
  \item{UN East Asia}
  \item{Country-specific (usually based on the latest WPP revision)}
  \item{Custom}
  }
}

