% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.nosocomial.infections}
\alias{dp.inputs.nosocomial.infections}
\title{Nosocomial infections among children}
\usage{
dp.inputs.nosocomial.infections(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
A data frame.
}
\description{
AIM takes as input numbers of nosocomial infections in children by five-year
age group and calendar year.
}
