% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.pmtct}
\alias{dp.inputs.pmtct}
\alias{dp.inputs.pmtct.retention.perinatal}
\alias{dp.inputs.pmtct.retention.postnatal}
\title{Get PMTCT inputs}
\usage{
dp.inputs.pmtct(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.inputs.pmtct.retention.perinatal(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.inputs.pmtct.retention.postnatal(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get PMTCT inputs
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.pmtct()}: Number or percent of women receiving PMTCT by year and regimen

\item \code{dp.inputs.pmtct.retention.perinatal()}: Percent retained on ART at delivery among HIV+ pregnant women

\item \code{dp.inputs.pmtct.retention.postnatal()}: Percentage of HIV+ breastfeeding mothers who interrupt PMTCT each month

}}
