% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.pop.percent}
\alias{dp.inputs.pop.percent}
\title{Get the percentage of the national population living in a subnational region}
\usage{
dp.inputs.pop.percent(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
population sizes by year.
}
\description{
Get the percentage of the national population living in a subnational region
}
\section{Details}{


  NOTE: THIS INPUT PERTAINS TO A FEATURE THAT HAS BEEN REMOVED FROM SPECTRUM.
  Countries preparing subnational projections could specify the subnational
  population by entering a national population size trend, and the share of
  the national population living in the subnational region.

  Use \code{dp.inputs.pop.country} to extract the national
  population of the country the subnational projection resides in.
}

