% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.shiny90.program}
\alias{dp.inputs.shiny90.program}
\alias{dp.inputs.shiny90.survey}
\title{Get inputs used to estimate knowledge of HIV status}
\usage{
dp.inputs.shiny90.program(dp.raw, direction = "wide")

dp.inputs.shiny90.survey(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A data frame

A data frame
}
\description{
Get inputs used to estimate knowledge of HIV status

Get inputs used to estimate knowledge of HIV status
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.shiny90.program()}: Data on HIV testing volumes and positive tests from programmatic reporting

\item \code{dp.inputs.shiny90.survey()}: Data on HIV testing history from population-based surveys

}}
