% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.survey.art.coverage}
\alias{dp.inputs.survey.art.coverage}
\title{Get ART coverage data entered from household surveys}
\usage{
dp.inputs.survey.art.coverage(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A data frame.
}
\description{
Get ART coverage data from household surveys that have been entered into
Spectrum for validation of its ART coverage estimates.
}
\section{Details}{


  Spectrum can store data from up to five household surveys. Data
  consist of ART coverage estimates by sex (male, female, both)
  and five-year age group (0-4, 5-9, ..., 75-79, 80+). Five statistics
  are stored for each group:

  \enumerate{
  \item{Number - not used}
  \item{Coverage - ART coverage point estimate}
  \item{Lower - 95\% confidence interval lower bound}
  \item{Upper - 95\% confidence interval upper bound}
  \item{N - sample size. Countries may enter non-integral survey-weighted sample sizes}
  }
}

