% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.use.external.pop}
\alias{dp.inputs.use.external.pop}
\title{Check whether an external population was specified}
\usage{
dp.inputs.use.external.pop(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored}
}
\value{
A list with two elements. Element \code{use.external.pop} is TRUE if
an external population was used, FALSE otherwise. Element \code{final.year}
specifies the final year of adjustments.
}
\description{
Spectrum users may enter externally-produced population targets by sex, age,
and year that can be used to align Spectrum population outputs with a
reference population projection. \code{dp.inputs.use.external.pop} indicates
whether this mechanism was used.
}
\section{Details}{


  Use of an external population file is optional and rare in Spectrum.
  Projections that use this mechanism may specify that adjustments to match
  external population targets are not done in every year, but only until a
  specified final year. Use \code{dp.inputs.external.pop} to access the
  external population sizes.
}

