% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.viral.suppression}
\alias{dp.inputs.viral.suppression}
\title{Get viral suppression inputs}
\usage{
dp.inputs.viral.suppression(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A data frame.
}
\description{
Get numbers of adults and children tested for viral suppression and numbers
whose test indicated viral suppression by year
}
