% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.output.bigpop}
\alias{dp.output.bigpop}
\title{Get Spectrum's calculated population}
\usage{
dp.output.bigpop(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get Spectrum's calculated population by age, sex, and year in long or wide
format
}
\section{Details}{


  If you plan to extract multiple, time-dependent DemProj variables, it is
  more efficient to save the first year and final year of the projection,
  then pass those values to subsequent function calls. If you leave the first
  and final year \code{NULL}, the code will call
  \code{dp.inputs.first.year()} and \code{dp.inputs.final.year()} each time
  you request a variable.
}

