% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.output.child.hiv.14}
\alias{dp.output.child.hiv.14}
\title{Get numbers of 14-year-olds living with HIV}
\usage{
dp.output.child.hiv.14(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get numbers of 14-year-olds living with HIV
}
\section{Details}{


  Spectrum stores detailed information about 14-year-olds living with HIV.
  This is used by EPP to account for the contribution of mother-to-child HIV
  transmission to HIV prevalence in younger adjults as children living with
  HIV reach adulthood. Data are stratified by sex, CD4 category, ART status,
  and timing of HIV acquisition (perinatal, breastfeeding at 0-6, 6-12, or
  12+ months of life)
}

