% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.output.csavr.deaths.hiv}
\alias{dp.output.csavr.deaths.hiv}
\alias{dp.output.csavr.incident.hiv}
\alias{dp.output.csavr.plhiv}
\title{CSAVR estimates}
\usage{
dp.output.csavr.deaths.hiv(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.output.csavr.incident.hiv(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.output.csavr.plhiv(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get estimates produced by Spectrum's Case Surveillance and Vital Registration (CSAVR) tool
}
\section{Functions}{
\itemize{
\item \code{dp.output.csavr.deaths.hiv()}: HIV-related deaths in adults

\item \code{dp.output.csavr.incident.hiv()}: New HIV infections in adults

\item \code{dp.output.csavr.plhiv()}: Adults living with HIV

}}
