% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.output.deaths.nonaids.excess}
\alias{dp.output.deaths.nonaids.excess}
\title{Get estimates of non-AIDS excess deaths}
\usage{
dp.output.deaths.nonaids.excess(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get estimates of non-AIDS excess deaths stratified by calendar year, sex
(male, female), age (0,1,...,79,80+), and ART status (on ART, not on ART)
}
