% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.output.pmtct}
\alias{dp.output.pmtct}
\alias{dp.output.pmtct.need}
\title{Get Spectrum PMTCT outputs}
\usage{
dp.output.pmtct(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.output.pmtct.need(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame
}
\description{
Get Spectrum PMTCT outputs
}
\section{Functions}{
\itemize{
\item \code{dp.output.pmtct()}: Number of pregnant women receiving PMTCT

\item \code{dp.output.pmtct.need()}: Number of pregnant women who need PMTCT

}}
