% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.output.ua.data}
\alias{dp.output.ua.data}
\title{Extract uncertainty analysis data}
\usage{
dp.output.ua.data(
  pjnz.file,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}

\item{direction}{Request "wide" (default) or "long" format data.}
}
\value{
A named list. See \code{Details} for a description of list contents.
}
\description{
Extract uncertainty analysis data
}
\section{Details}{

\code{dp.output.ua.data} returns data used to calculate uncertainty bounds
on AIM estimates along with metadata captured when Spectrum's Uncertainty
Analysis (UA) tool was run. The data and metadata items consist of:
\describe{
\item{Version}{The version number of the file that stores UA data.}
\item{Date}{The date and time the UA tool was run, according to the
user's system clock. Note that SpectrumUtils interprets this in UTC timezone,
but Spectrum does not save the timezone so the date is only accurate to
within 24 hours.}
\item{AIDSDeaths}{The number of HIV-related deaths in the final year
of the projection across all ages and both sexes. Spectrum considers UA
results valid only if this number of deaths agrees with the point estimate
of HIV-related deaths. If these disagree, the user probably changed some
model inputs since the last time UA was run.}
\item{Data}{Uncertainty analysis bounds by sex and year for several
key indicators. Bounds are expressed as a ratio relative to the point
estimate. For example, a high bound of 1.1 means that the upper bound
is 10\% higher than the point estimate. Point estimates for indicators
must be extracted separately using appropriate \code{dp.extract}
functions.}
}
}

