% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.geo.info}
\alias{extract.geo.info}
\title{Read Spectrum geographic metadata}
\usage{
extract.geo.info(pjn.raw)
}
\arguments{
\item{pjn.raw}{Raw projection parameter data, as returned by
\code{read.raw.pjn()}}
}
\value{
A data frame with three elements:

  \code{iso.code} ISO-3166 numeric country code

  \code{snu.name} Subnational unit name. This is empty for national
  projections

  \code{snu.code} Subnational numeric code. This is 0 for national
  projections
}
\description{
Extract the country ISO-3166 numeric code and subnational unit name and id
number from Spectrum projection data
}
\section{Details}{


  Numeric subnational codes were assigned by Avenir Health for internal use
  in Spectrum. Unlike national ISO-3166 codes, these subnational codes are
  not governed by any international organization and may not match codes used
  for similar purposes by other organizations.
}

