% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.proj.name}
\alias{extract.proj.name}
\title{Read the projection name}
\usage{
extract.proj.name(pjn.raw)
}
\arguments{
\item{pjn.raw}{Raw projection parameter data, as returned by
\code{read.raw.pjn()}}
}
\value{
The projection name. This will usually match the PJNZ file name. This
  may not be true if the PJNZ is renamed manually, such as by renaming it in
  Windows Explorer.
}
\description{
Read the projection name
}
