% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-ha.R
\name{ha.inputs.condom.use}
\alias{ha.inputs.condom.use}
\alias{ha.inputs.viral.suppression}
\title{HIV program coverage inputs}
\usage{
ha.inputs.condom.use(
  ha.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

ha.inputs.viral.suppression(
  ha.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{ha.raw}{Goals ASM module data in raw format, as returned by
\code{read.raw.ha()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{ha.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{ha.inputs.final.year()}}
}
\value{
a data frame
}
\description{
HIV program coverage inputs
}
\section{Functions}{
\itemize{
\item \code{ha.inputs.condom.use()}: Condom use inputs by year and age group

\item \code{ha.inputs.viral.suppression()}: Input levels of viral suppression on ART by year, sex, and age group

}}
