% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-ha.R
\name{ha.inputs.fitdata.trend}
\alias{ha.inputs.fitdata.trend}
\title{Data for model calibration}
\usage{
ha.inputs.fitdata.trend(
  ha.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{ha.raw}{Goals ASM module data in raw format, as returned by
\code{read.raw.ha()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
a data frame.
}
\description{
Data for model calibration
}
\section{Functions}{
\itemize{
\item \code{ha.inputs.fitdata.trend()}: HIV incidence or prevalence data points

}}
