% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-ha.R
\name{ha.inputs.transmission}
\alias{ha.inputs.transmission}
\title{Extract inputs listed in the Goals ASM "Epidemiology" form}
\usage{
ha.inputs.transmission(ha.raw, direction = "wide")
}
\arguments{
\item{ha.raw}{Goals ASM module data in raw format, as returned by
\code{read.raw.ha()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{ha.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{ha.inputs.final.year()}}
}
\value{
a data frame
}
\description{
Extract transmission probabilities by HIV-positive partner's sex
}
