% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-hv.R
\name{hv.inputs.first.year}
\alias{hv.inputs.first.year}
\alias{hv.inputs.final.year}
\title{Spectrum projection time span}
\usage{
hv.inputs.first.year(hv.raw, direction = "wide")

hv.inputs.final.year(hv.raw, direction = "wide")
}
\arguments{
\item{hv.raw}{Goals module data in raw format, as returned by
\code{read.raw.hv()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
the requested year
}
\description{
Spectrum projection time span
}
\details{
This can also be accessed from the DemProj module via \code{dp.inputs.first.year}.
}
\section{Functions}{
\itemize{
\item \code{hv.inputs.first.year()}: First year of the projection

\item \code{hv.inputs.final.year()}: Final year of the projection

}}
