% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-hv.R
\name{hv.output.adults}
\alias{hv.output.adults}
\title{Get the estimated numbers of adults ages 15-49 in each model compartment over
time.}
\usage{
hv.output.adults(
  hv.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{hv.raw}{Goals module data in raw format, as returned by
\code{read.raw.hv()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{hv.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{hv.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get the estimated numbers of adults ages 15-49 in each model compartment over
time.
}
