% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.module.data}
\alias{read.module.data}
\title{Load Spectrum module data}
\usage{
read.module.data(pjnz.file, extension = "DP")
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}

\item{extension}{Module data file name extension}
}
\value{
an unformatted table of module data
}
\description{
Read Spectrum module data as an unformatted table
}
\examples{
dp.data = read.module.data("Antarctica.PJNZ", extension="DP")
}
