% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{read.raw.dp}
\alias{read.raw.dp}
\title{Load Spectrum DemProj module data}
\usage{
read.raw.dp(pjnz.file)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}
}
\value{
an unformatted table of module data
}
\description{
Read Spectrum DemProj module data as an unformatted table
}
\examples{
dp.data = read.data.dp("Antarctica.PJNZ")
}
