% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-ha.R
\name{read.raw.ha}
\alias{read.raw.ha}
\title{Load Spectrum Goal ASM module data}
\usage{
read.raw.ha(pjnz.file)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}
}
\value{
an unformatted table of module data
}
\description{
Read Spectrum Goals ASM module data as an unformatted table
}
\examples{
ha.data = read.data.ha("Antarctica.PJNZ")
}
