% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-hv.R
\name{read.raw.hv}
\alias{read.raw.hv}
\title{Load Spectrum Goals RSM module data}
\usage{
read.raw.hv(pjnz.file)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}
}
\value{
an unformatted table of module data
}
\description{
Read Spectrum Goals Risk-Structured Model (RSM) module data as an unformatted table
}
\examples{
hv.data = read.data.hv("Antarctica.PJNZ")
}
