% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.raw.pjn}
\alias{read.raw.pjn}
\title{Read Spectrum projection parameters}
\usage{
read.raw.pjn(pjnz.file)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}
}
\value{
an unformatted table of projection data
}
\description{
Read Spectrum projection parameters as an unformatted table
}
\examples{
dp.data = read.raw.pjn("Antarctica.PJNZ")
}
