% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-rn.R
\name{read.raw.rn}
\alias{read.raw.rn}
\title{Load Spectrum Resource Needs module data}
\usage{
read.raw.rn(pjnz.file)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}
}
\value{
an unformatted table of module data
}
\description{
Read Spectrum Resource Needs module data as an unformatted table
}
\examples{
rn.data = read.data.rn("Antarctica.PJNZ")
}
