% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-spu.R
\name{read.raw.spu}
\alias{read.raw.spu}
\title{Read HIV uncertainty data}
\usage{
read.raw.spu(pjnz.file)
}
\arguments{
\item{pjnz.file}{The Spectrum file to extract data from}
}
\value{
a dataframe with one row per row in the SPU file
}
\description{
Uncertainty around HIV incidence and prevalence for AIM is sampled using
tools like EPP and CSAVR. These tools produce store the uncertainty data in
an SPU file within a projection. An SPU file is a CSV file that stores
coupled estimates of HIV incidence and prevalence that correspond to
underlying sampled model parameter values. \code{read.raw.spu} loads the
contents of this file with minimal processing. Other functions, like
\code{format.spu} can be used to reformat these data for easier use.
}
\examples{
dp.data = read.data.spu("Antarctica.PJNZ")
}
