% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-rn.R
\name{rn.inputs.coverage}
\alias{rn.inputs.coverage}
\title{Resource needs module (RNM) coverage inputs inputs}
\usage{
rn.inputs.coverage(rn.raw, direction = "wide", first.year, final.year)
}
\arguments{
\item{rn.raw}{RNM module data in raw format, as returned by
\code{read.raw.rn()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection.}

\item{final.year}{Final year of the projection.}
}
\value{
A data frame.
}
\description{
Resource needs module (RNM) coverage inputs inputs
}
