% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-spu.R
\name{spu.indicator.data}
\alias{spu.indicator.data}
\title{Reformat SPU data as a data frame}
\usage{
spu.indicator.data(raw.spu, direction = "wide")
}
\arguments{
\item{raw.spu}{SPU data returned by \code{read.raw.spu}}
}
\value{
a dataframe in long or wide format
}
\description{
Return a data frame of HIV incidence and prevalence estimates. Uncertainty
data in SPU files are sampled from some distribution. Sampling may be done
with replacement. Formatted SPU data include a sample id column ("Sample"),
the sampling frequency ("Frequency"), and Year. If "wide" output is
requested, the dataframe will have columns for HIV prevalence and incidence;
if "long" output is requested, the dataframe will have an Indicator and Value
columns instead. The SPU data includes the best-fitting incidence and
prevalence point estimate. This has sample ID 0 and missing (NA) sampling
frequency.
}
\examples{
spu = read.raw.spu("Antarctica.PJNZ")
dat = spu.indicator.data(spu, direction="long")
}
