% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.first.year}
\alias{dp.inputs.first.year}
\title{Get the first year of a Spectrum projection}
\usage{
dp.inputs.first.year(dp.raw, direction = "wide")
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
the first year of the projection.
}
\description{
Get the first year of a Spectrum projection
}
