% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-dp.R
\name{dp.inputs.viral.suppression}
\alias{dp.inputs.viral.suppression}
\alias{dp.inputs.viral.suppression.threshold}
\title{Get viral suppression inputs}
\usage{
dp.inputs.viral.suppression(
  dp.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

dp.inputs.viral.suppression.threshold(
  dp,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{dp.raw}{DemProj module data in raw format, as returned by
\code{read.raw.dp()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{dp.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{dp.inputs.final.year()}}
}
\value{
A data frame.
}
\description{
Get viral suppression inputs
}
\details{
Some versions of Spectrum only allowed viral suppression data to be
  entered during 2010-2025. Data are not available in earlier or later years
  in files saved with these versions. Spectrum 6.29 and later can save data
  for every year in a projection, though countries may not input anything in
  most years.
}
\section{Functions}{
\itemize{
\item \code{dp.inputs.viral.suppression()}: Numbers of people who had a viral load measured and numbers who were virally suppressed

\item \code{dp.inputs.viral.suppression.threshold()}: The threshold (in virus copies/mL) used to classify people as virally suppressed

}}
