% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract.raw.tag}
\alias{extract.raw.tag}
\title{Extract specific Spectrum data}
\usage{
extract.raw.tag(mod.raw, tag, fmt)
}
\arguments{
\item{mod.raw}{Raw module data returned by read.module.data or
module-specific functions}

\item{tag}{A module variable name}

\item{fmt}{Format parameters. See "Details"}
}
\value{
A matrix storing the module variable values.
}
\description{
Read raw Spectrum data from a given module
}
\section{Details}{


  Spectrum module data uses either "modvar" or non-modvar formats.

  modvar formats group data into blocks delimited by a starting tag (e.g.
  <FirstYear MV2>) and matching "<End>" tag. Starting tags have an embedded
  version number (e.g. "MV2"; "MV" for version 1) to accommodate the
  evolution of Spectrum file formats over time.

  non-modvar formats also include a starting tag (e.g. "<Projection Name>")
  but no embedded version number, and no matching "<End>" tag.

  The \code{fmt} parameter is a named list that specifies the module data
  format and variable layout that should contain the following:

  \code{fmt$is.modvar} is \code{TRUE} for modvar format modules and
  \code{FALSE} otherwise.

  \code{fmt$offset}: the number of rows between the starting tag and the
  first row of data.

  \code{fmt$nrow}: Number of rows of data, relative to the starting tag row
  number + \code{fmt$offset}

  \code{fmt$ncol}: Number of columns of data, relative to mod.raw$Data.

  \code{fmt$cast} is a function used to cast the data from a string to a
  desired type. For example, set \code{fmt$type=as.numeric} to return numeric
  data.
}

