% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-ha.R
\name{ha.inputs.lifetime.partners}
\alias{ha.inputs.lifetime.partners}
\alias{ha.inputs.partner.age}
\alias{ha.inputs.partner.age.par}
\alias{ha.inputs.mixing.age}
\title{Extract Goals ASM behavioral inputs}
\usage{
ha.inputs.lifetime.partners(
  ha.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)

ha.inputs.partner.age(ha.raw, direction = "wide")

ha.inputs.partner.age.par(ha.raw, direction = "wide")

ha.inputs.mixing.age(ha.raw, direction = "wide")
}
\arguments{
\item{ha.raw}{Goals ASM module data in raw format, as returned by
\code{read.raw.ha()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{ha.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{ha.inputs.final.year()}}
}
\value{
a data frame
}
\description{
Extract Goals ASM behavioral inputs
}
\section{Functions}{
\itemize{
\item \code{ha.inputs.lifetime.partners()}: Time trend in numbers of lifetime partners

\item \code{ha.inputs.partner.age()}: Partner change rates by age in the last year of projection

\item \code{ha.inputs.partner.age.par()}: Parameters describing relative rates of partner change by age

\item \code{ha.inputs.mixing.age()}: Mixing preferences by age

}}
\section{Details}{


\itemize{
\item{\code{ha.inputs.partner.age.par} Returns the peak and median in relative rates of
partner change by age after sexual debut. Relative rates of partner change at age a are
calculated as dlnorm(a+0.5-debut, logmu, logsd) where logmu = median-debut and logsd=sqrt(logmu - ln(peak - debut))}
\item{\code{ha.inputs.mixing.age}} Returns mixing preferences of females by male age. In wide format, rows correspond to male age, columns to female age.
}
}

