% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-ha.R
\name{ha.output.population.reached}
\alias{ha.output.population.reached}
\title{Get the number of people reached by HIV interventions}
\usage{
ha.output.population.reached(
  ha.raw,
  direction = "wide",
  first.year = NULL,
  final.year = NULL
)
}
\arguments{
\item{ha.raw}{Goals ASM module data in raw format, as returned by
\code{read.raw.ha()}}

\item{direction}{Request "wide" (default) or "long" format data.}

\item{first.year}{First year of the projection. If \code{first.year=NULL}, it
will be filled in using \code{ha.inputs.first.year()}}

\item{final.year}{Final year of the projection. If \code{final.year=NULL}, it
will be filled in using \code{ha.inputs.final.year()}}
}
\value{
a data frame
}
\description{
Get the number of people reached by HIV interventions
}
