% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-hv.R
\name{hv.inputs.calibration.data}
\alias{hv.inputs.calibration.data}
\title{Extract survey and study data used for model fitting}
\usage{
hv.inputs.calibration.data(hv.raw, direction = "wide")
}
\arguments{
\item{hv.raw}{Goals module data in raw format, as returned by
\code{read.raw.hv()}}

\item{direction}{Ignored; included for compatibility with similar functions.}
}
\value{
A data frame.
}
\description{
Extract survey and study data used for model fitting
}
